Fixed issue caused by manually changing column types in second-previous commit
Issue was caused when the column types of several IdentityUser tables were manually changed inside the migration's source code. This caused Add-Migration to want to put revert back the changes.

In this commit, the initial migration has been made fully generalised, such that, if an Add-Migration is done for both cases of UseMySQL = true or false, then the added migration will be empty.